Rename time.rs to parse_time.rs
Conflicts with crate time.
