+++ src/sleep/sleep.rs
@@ -20,8 +20,8 @@ use std::u32::MAX as U32_MAX;
 #[macro_use]
 mod util;
 
-#[path = "../common/time.rs"]
-mod time;
+#[path = "../common/parse_time.rs"]
+mod parse_time;
 
 static NAME: &'static str = "sleep";
 static VERSION: &'static str = "1.0.0";
@@ -67,7 +67,7 @@ specified by the sum of their values.", NAME, VERSION);
 
 fn sleep(args: Vec<String>) {
     let sleep_time = args.iter().fold(0.0, |result, arg|
-        match time::from_str(&arg[..]) {
+        match parse_time::from_str(&arg[..]) {
             Ok(m) => m + result,
             Err(f) => crash!(1, "{}", f),
         });
